move Job Template code into its own file
